Move modal style to theme file
