+++ frontend/src/utils/utils.ts
@@ -19,16 +19,3 @@ export const post = (body: RequestBody) => {
     body: JSON.stringify(body),
   };
 };
-
-export const modalStyle = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 300,
-  bgcolor: 'background.paper',
-  color: 'black',
-  borderRadius: '10px',
-  p: 4,
-  textAlign: 'center',
-};
